final touches on benchmarks
had to move the utils into the crate, because otherwise it would be
picked up by the bechmarks and fail because libtest would again complain
about the --output-format argument
