+++ benchmarks/Cargo.toml
@@ -13,12 +13,12 @@ publish = false
 bench = false # so libtest doesn't eat the arguments for criterion
 
 [dependencies]
+rand = { version = "0.8" }
 
 [dev-dependencies]
 libcrux = { path = "../", features = ["rand", "tests"] }
 libcrux-kem = { path = "../libcrux-kem", features = ["tests"] }
 libcrux-ml-kem = { path = "../libcrux-ml-kem" }
-rand = { version = "0.8" }
 rand_core = { version = "0.6" }
 # Benchmarking "RustCrypto"
 chacha20poly1305 = "0.10"
